+++ pkg/scalers/external_scaler.go
@@ -151,15 +151,25 @@ func (s *externalScaler) GetMetricSpecForScaling(ctx context.Context) []v2.Metri
 	}
 
 	for _, spec := range response.MetricSpecs {
+		metricType := s.metricType
+		if spec.MetricType != nil {
+			switch *spec.MetricType {
+			case string(v2.ValueMetricType), string(v2.AverageValueMetricType):
+				metricType = v2.MetricTargetType(*spec.MetricType)
+			default:
+				s.logger.Info("Invalid MetricType received from external scaler, falling back to default", "MetricType", *spec.MetricType)
+			}
+		}
+
 		externalMetric := &v2.ExternalMetricSource{
 			Metric: v2.MetricIdentifier{
 				Name: GenerateMetricNameWithIndex(s.metadata.triggerIndex, spec.MetricName),
 			},
 		}
 		if spec.TargetSizeFloat > 0 {
-			externalMetric.Target = GetMetricTargetMili(s.metricType, spec.TargetSizeFloat)
+			externalMetric.Target = GetMetricTargetMili(metricType, spec.TargetSizeFloat)
 		} else {
-			externalMetric.Target = GetMetricTarget(s.metricType, spec.TargetSize)
+			externalMetric.Target = GetMetricTarget(metricType, spec.TargetSize)
 		}
 
 		// Create the metric spec for the HPA